[Chore]: 인디케이터를 적용한 뷰 분리
- 공통 레이아웃 컴포넌트 생성(InfoStepContainer)
